feat: rename http-client to fetch-client
