+++ apps/react-demo/src/shared/app-fetch-client.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@ns3/di';
-import { FetchClient } from '@ns3/http-client';
+import { FetchClient } from '@ns3/fetch-client';
 import { fetchClientFactory } from './app-fetch-client-utils';
 import { DelayInterceptor } from './delay.interceptor';
