+++ libs/fetch-client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@ns3/http-client",
+  "name": "@ns3/fetch-client",
   "version": "0.0.1",
   "license": "MIT",
   "sideEffects": false,
