+++ workspace.json
@@ -2,7 +2,7 @@
   "version": 2,
   "projects": {
     "di": "libs/di",
-    "http-client": "libs/http-client",
+    "fetch-client": "libs/fetch-client",
     "react-demo": "apps/react-demo",
     "react-di": "libs/react-di",
     "react-utils": "libs/react-utils",
